ai-service/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
